+++ apps/probability/calculation/right_integral_calculation.cpp
@@ -1,7 +1,7 @@
 #include "right_integral_calculation.h"
 #include <assert.h>
 #include <ion.h>
-#include <math.h>
+#include <cmath>
 
 namespace Probability {
 
@@ -31,7 +31,7 @@ I18n::Message RightIntegralCalculation::legendForParameterAtIndex(int index) {
 
 void RightIntegralCalculation::setParameterAtIndex(float f, int index) {
   assert(index >= 0 && index < 2);
-  float rf = roundf(f/k_precision)*k_precision;
+  float rf = std::round(f/k_precision)*k_precision;
   if (index == 0) {
     m_lowerBound = rf;
   }
@@ -60,10 +60,10 @@ void RightIntegralCalculation::compute(int indexKnownElement) {
   if (indexKnownElement == 0) {
     m_result = m_law->rightIntegralFromAbscissa(m_lowerBound);
     /* Results in probability application are rounder to 3 decimals */
-    m_result = roundf(m_result/k_precision)*k_precision;
+    m_result = std::round(m_result/k_precision)*k_precision;
   } else {
     m_lowerBound = m_law->rightIntegralInverseForProbability(&m_result);
-    m_lowerBound = roundf(m_lowerBound/k_precision)*k_precision;
+    m_lowerBound = std::round(m_lowerBound/k_precision)*k_precision;
   }
 }
