+++ apps/sequence/graph/go_to_parameter_controller.cpp
@@ -1,12 +1,13 @@
 #include "go_to_parameter_controller.h"
 #include "../app.h"
 #include <assert.h>
+#include <cmath>
 
 namespace Sequence {
 
 bool GoToParameterController::setParameterAtIndex(int parameterIndex, float f) {
   assert(parameterIndex == 0);
-  return Shared::FunctionGoToParameterController::setParameterAtIndex(parameterIndex, roundf(f));
+  return Shared::FunctionGoToParameterController::setParameterAtIndex(parameterIndex, std::round(f));
 }
 
 }
