+++ apps/sequence/values/interval_parameter_controller.cpp
@@ -1,4 +1,5 @@
 #include "interval_parameter_controller.h"
+#include <cmath>
 
 using namespace Shared;
 
@@ -24,7 +25,7 @@ bool IntervalParameterController::setParameterAtIndex(int parameterIndex, float 
     app()->displayWarning(I18n::Message::ForbiddenValue);
     return false;
   }
-  float parameter = roundf(f);
+  float parameter = std::round(f);
   if (parameterIndex == 2 && parameter == 0.0f) {
     parameter = 1.0f;
   }
